chore(api-gateway): project init — add env example, dependencies and entrypoints
