+++ backend/api_gateway/.env.example
@@ -1,8 +1,8 @@
-SECRET=
-PORT=
-MS_AUTH_URL=
-MS_CUSTOMER_URL=
-MS_EMPLOYEE_URL=
-MS_FLIGHTS_URL=
-MS_RESERVE_URL=
-SAGAS_URL=
+SECRET= changemechangemechangemechangeme12
+PORT= 8080
+MS_AUTH_PORT=8081
+MS_CUSTOMER_PORT=8082
+MS_EMPLOYEE_PORT=8083
+MS_FLIGHT_PORT=8084
+MS_RESERVE_PORT=8085
+SAGA_PORT=8087
