+++ backend/api_gateway/package.json
@@ -1,7 +1,13 @@
 {
+  "scripts": {
+    "start": "node index.js",
+    "dev": "nodemon index.js"
+  },
   "dependencies": {
+    "axios": "^1.8.4",
     "body-parser": "^2.2.0",
     "cookie-parser": "^1.4.7",
+    "dotenv": "^16.5.0",
     "dotenv-safe": "^9.1.0",
     "express": "^5.1.0",
     "express-http-proxy": "^2.1.1",
